Splitted map checker in multiple classes
