+++ jsettlers.mapcreator/src/jsettlers/mapcreator/localization/labels_de.properties
@@ -174,5 +174,10 @@ tools.category.soldier        = Krieger
 tools.category.specialist     = Spezialisten
 tools.category.worker         = Arbeiter
 
+validation.at-position           = Bei Position %d/%d
+validation.blockingborder.header = Kartenr\u00E4nder m\u00FCssen blockieren
+validation.resource.header       = Flascher untergrund
+validation.resource.text         = "%s" darf kein(e) "%s" haben
+
 window.current-player = Setze Objekte f\u00FCr Spieler:
 window.header         = %s - JSettlers Karteneditor
