+++ jsettlers.mapcreator/src/jsettlers/mapcreator/localization/labels_en_US.properties
@@ -174,5 +174,10 @@ tools.category.soldier        = soldier
 tools.category.specialist     = specialist
 tools.category.worker         = worker
 
+validation.at-position           = At position %d/%d
+validation.blockingborder.header = All border positions must be blocking
+validation.resource.header       = Wrong landscape
+validation.resource.text         = "%s" may not have "%s"
+
 window.current-player = set objects for:
 window.header         = %s - JSettlers Map Editor
