+++ jsettlers.mapcreator/src/jsettlers/mapcreator/mapvalidator/ValidatorRunnable.java
@@ -8,7 +8,12 @@ import javax.swing.SwingUtilities;
 import jsettlers.mapcreator.data.MapData;
 import jsettlers.mapcreator.mapvalidator.result.ValidationList;
 import jsettlers.mapcreator.mapvalidator.tasks.AbstractValidationTask;
-import jsettlers.mapcreator.mapvalidator.tasks.ValidateAll;
+import jsettlers.mapcreator.mapvalidator.tasks.ValidateBlockingBorderPositions;
+import jsettlers.mapcreator.mapvalidator.tasks.ValidateBuildings;
+import jsettlers.mapcreator.mapvalidator.tasks.ValidateDrawBuildingCircle;
+import jsettlers.mapcreator.mapvalidator.tasks.ValidateLandscape;
+import jsettlers.mapcreator.mapvalidator.tasks.ValidatePlayerStartPosition;
+import jsettlers.mapcreator.mapvalidator.tasks.ValidateResources;
 
 /**
  * The validation runnable running in the thread queue
@@ -38,6 +43,16 @@ public class ValidatorRunnable implements Runnable {
 	private final MapData data;
 
 	/**
+	 * Player data array
+	 */
+	protected byte[][] players;
+
+	/**
+	 * Border array
+	 */
+	protected boolean[][] borders;
+
+	/**
 	 * Constructor
 	 * 
 	 * @param resultListener
@@ -49,7 +64,27 @@ public class ValidatorRunnable implements Runnable {
 		this.resultListener = resultListener;
 		this.data = data;
 
-		registerTask(new ValidateAll());
+		// keep order, will be executed in this order
+		registerTask(new ValidateBlockingBorderPositions());
+		registerTask(new ValidateDrawBuildingCircle());
+		registerTask(new ValidateBuildings());
+		registerTask(new ValidateLandscape());
+		registerTask(new ValidateResources());
+		registerTask(new ValidatePlayerStartPosition());
+	}
+
+	/**
+	 * Initialize the player array
+	 */
+	private void initPlayerData() {
+		players = new byte[data.getWidth()][data.getHeight()];
+		for (int x = 0; x < data.getWidth(); x++) {
+			for (int y = 0; y < data.getHeight(); y++) {
+				players[x][y] = (byte) -1;
+			}
+		}
+
+		borders = new boolean[data.getWidth()][data.getHeight()];
 	}
 
 	/**
@@ -62,16 +97,23 @@ public class ValidatorRunnable implements Runnable {
 		tasks.add(task);
 		task.setData(data);
 		task.setList(list);
+		task.setPlayers(players);
+		task.setBorders(borders);
 	}
 
 	@Override
 	public void run() {
-		list.addHeader("REAL");
+		initPlayerData();
 
 		for (AbstractValidationTask t : tasks) {
 			t.doTest();
 		}
 
+		data.setPlayers(players);
+		data.setBorders(borders);
+		data.setFailpoints(new boolean[data.getWidth()][data.getHeight()]);
+
+		// fire result to UI
 		list.prepareToDisplay();
 		fireResult();
 	}
